EXECUTIVE_AI_ASSISTANT/LOGOS-ECOSYSTEM-VERSION-BETA.001/backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.25.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Cache
redis==5.0.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# AI/ML
# anthropic==0.8.1  # Removed - using generic AI integration
openai==1.8.0
langchain==0.1.0
tiktoken==0.5.2

# Vector Databases
pinecone-client==3.0.0
weaviate-client==4.4.0
qdrant-client==1.7.0
chromadb==0.4.22
faiss-cpu==1.7.4

# Embeddings & NLP
sentence-transformers==2.3.1
cohere==4.37
nltk==3.8.1
spacy==3.7.2

# ML Frameworks
torch==2.1.2
transformers==4.36.2
datasets==2.16.1
accelerate==0.25.0
optimum==1.16.0

# Search & Ranking
whoosh==2.7.4
rank-bm25==0.2.2
elasticsearch==8.11.1

# ML Utilities
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.0

# ML Tracking & Optimization
wandb==0.16.2
optuna==3.5.0
mlflow==2.9.2

# Graph Processing
networkx==3.2.1

# Additional AI Tools
huggingface-hub==0.20.1
safetensors==0.4.1
einops==0.7.0
bitsandbytes==0.41.3

# Monitoring
prometheus-client==0.19.0
python-json-logger==2.0.7

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Development
black==23.12.1
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.0

# System monitoring
psutil==5.9.7

# Additional utilities
pendulum==3.0.0
httpx==0.26.0
aiofiles==23.2.1

# Whitelabel Dependencies
cssutils==2.9.0
libsass==0.23.0
dnspython==2.4.2
certbot==2.8.0
certbot-nginx==2.8.0

# Audio processing
gtts==2.3.0
SpeechRecognition==3.9.0
pygame==2.1.3
pydub==0.25.1
openai-whisper==20230314
soundfile==0.12.1
librosa==0.10.1
scipy==1.11.4
pyaudio==0.2.14
azure-cognitiveservices-speech==1.34.0
boto3==1.34.14
google-cloud-texttospeech==2.16.1
google-cloud-speech==2.23.0
sounddevice==0.4.6
torchaudio==2.1.2
joblib==1.3.2

# Automotive Integration
obd==0.7.1
python-can==4.3.0
pybluez==0.23
pycarwings2==2.14
tesla-api==3.0.5
smartcar==6.0.0
paho-mqtt==2.0.0
websockets==12.0
protobuf==4.25.2
mercedes-me-api==0.11.1
bmw-connected-drive==0.7.0
fordpass==0.0.5
hyundai-kia-connect-api==3.0.1
volvo-api==0.1.0

# IoT Protocols
aiocoap==0.4.7
paho-mqtt==2.0.0
asyncio-mqtt==0.16.2
pymodbus==3.5.4
minimalmodbus==2.1.0
pybluez==0.23
bleak==0.21.1
aioblescan==0.2.14
zigpy==0.60.2
zigpy-znp==0.11.6
zigpy-deconz==0.21.1
bellows==0.37.3
pyserial==3.5
RPi.GPIO==0.7.1
pyLoRa==0.3.0
sx127x==0.1.0
cbor2==5.5.1
scapy==2.5.0

# Payment Processing
stripe==7.8.0
paypalserversdk==0.5.0
web3==6.13.0
eth-account==0.10.0
bitcoin==1.1.42
bitcoinlib==0.6.14
blockcypher==1.0.93
stellar-sdk==9.0.0
py-algorand-sdk==2.4.0
solana==0.30.2
ccxt==4.2.7
python-binance==1.0.19
coinbase-commerce==1.0.1
qrcode==7.4.2
Pillow==10.2.0